Changed double to decimal

+++ CryptoLab.Domain/Domain/Wallet.cs
@@ -6,14 +6,14 @@ namespace CryptoLab.Domain.Domain
     {
         public Guid Id { get; protected set; }
         public string Currnecy { get; protected set; }
-        public double AmountOfMoney { get; protected set; }
+        public decimal AmountOfMoney { get; protected set; }
         public DateTime UpdatedAt { get; protected set; }
         public Guid UserId { get; protected set; }
         public User User { get; protected set; }
         
 
 
-        public Wallet(string currency, double amountOfMoney, User user)
+        public Wallet(string currency, decimal amountOfMoney, User user)
         {
             Id = Guid.NewGuid();
             SetCurrnecy(currency);
@@ -32,9 +32,9 @@ namespace CryptoLab.Domain.Domain
             Currnecy = currency.ToUpperInvariant();
         }
 
-        public void SetAmountOfMoney(double amountOfMoney)
+        public void SetAmountOfMoney(decimal amountOfMoney)
         {
-            if(amountOfMoney < 0.0)
+            if(amountOfMoney < 0.0m)
                 throw new Exception("Amount of money can not be null");
 
             AmountOfMoney = amountOfMoney;
