+++ CryptoLab.Infrastructure/IServices/ITransactionService.cs
@@ -5,7 +5,7 @@ namespace CryptoLab.Infrastructure.IServices
 {
     public interface ITransactionService
     {
-         Task FastBuyTransactionAsync(string toCurrnecy, double amount, Guid userId);
-         Task FastSellTransactionAsync(string fromCurrnecy, double amount, Guid userId);
+         Task FastBuyTransactionAsync(string toCurrnecy, decimal amount, Guid userId);
+         Task FastSellTransactionAsync(string fromCurrnecy, decimal amount, Guid userId);
     }
 }
